% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_cluster.R
\name{characterize_cluster}
\alias{characterize_cluster}
\title{Title}
\usage{
characterize_cluster(
  .data = NULL,
  gran1 = NULL,
  gran2 = NULL,
  harmony_tbl = NULL,
  nperm = 2,
  nsamp = 2,
  method = "dist_clust",
  response = NULL,
  quantile_prob_val = c(0.1, 0.25, 0.5, 0.75, 0.9),
  kopt = NULL,
  key = NULL
)
}
\arguments{
\item{quantile_prob_val}{}
}
\value{

}
\description{
Title
}
\examples{
library(gravitas)
library(tidyverse)
library(parallel)
library(tsibble)
library(rlang)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
  harmonies <- sm \%>\%
  harmony(
  ugran = "month",
  filter_in = "wknd_wday",
  filter_out = c("hhour", "fortnight")
  )
v = suppressWarnings(characterize_cluster(sm, gran1 = "hour_day", harmony_tbl = harmonies, method = "scaled_dist_clust"))
v
v1 = suppressWarnings(characterize_cluster(sm, gran1 = "hour_day", harmony_tbl = harmonies, method = "wpd_clust"))
v1
v2 = suppressWarnings(characterize_cluster(sm, gran1 = "hour_day", harmony_tbl = harmonies, method = "dist_clust", kopt = 2))
v2
}
