% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_gran.R
\name{clust_gran}
\alias{clust_gran}
\title{compute distances and groups from algorithm based on raw distributions}
\usage{
clust_gran(dist, kopt = NULL, method = "ward.D")
}
\arguments{
\item{dist}{a dissimilarity structure as produced by dist.}

\item{kopt}{the number of clusters}

\item{method}{method in stats::hclust()}
}
\value{

}
\description{
compute distances and groups from algorithm based on raw distributions
}
\examples{
library(gravitas)
library(tidyverse)
library(tsibble)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
response = "general_supply_kwh"
sm \%>\% scale_gran(method = "robust") \%>\% dist_gran(gran1 = "hour_day")\%>\%
clust_gran("hour_day")
sm \%>\% scale_gran(method = "nqt") \%>\% dist_gran(gran1 = "hour_day")\%>\%
clust_gran()
dist_wpd(sm, harmony_tbl = h)) \%>\% clust_gran()
}
