% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_algo.R
\name{compute_wpd}
\alias{compute_wpd}
\title{Title}
\usage{
compute_wpd(
  .data = NULL,
  gran1 = NULL,
  gran2 = NULL,
  harmony_tbl = NULL,
  nperm = 2,
  nsamp = 2,
  method = "dist_clust",
  response = NULL,
  quantile_prob_val = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{quantile_prob_val}{}
}
\value{

}
\description{
Title
}
