% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cust_by_cat.R
\name{cust_by_cat}
\alias{cust_by_cat}
\title{Title}
\usage{
cust_by_cat(.data, gran1 = NULL, gran2 = NULL, response = NULL)
}
\arguments{
\item{response}{}
}
\value{

}
\description{
Title
}
\examples{
library(gravitas)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
v1 = cust_by_cat(sm, "hour_day")
v1
}
