% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_plot.R
\name{den_plot}
\alias{den_plot}
\title{Title}
\usage{
den_plot(
  .data,
  gran1 = NULL,
  gran2 = NULL,
  response = NULL,
  quantile_prob_val = seq(0.1, 0.9, 0.1),
  kopt = NULL
)
}
\arguments{
\item{kopt}{}
}
\value{

}
\description{
Title
}
\examples{
library(gravitas)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936", "10018064"))
gran1 = "hour_day"
gran2 = NULL
suppressWarnings(den_plot(sm, "hour_day", kopt = 3))
}
