% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clust.R
\name{dist_clust}
\alias{dist_clust}
\title{compute distances and groups from algorithm based on raw distributions}
\usage{
dist_clust(
  .data,
  gran1 = NULL,
  gran2 = NULL,
  response = NULL,
  quantile_prob_val = seq(0.1, 0.9, 0.1),
  kopt = NULL,
  key = NULL
)
}
\arguments{
\item{quantile_prob_val}{}
}
\value{

}
\description{
compute distances and groups from algorithm based on raw distributions
}
\examples{
library(gravitas)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
v1 = suppressWarnings(dist_clust(sm, "hour_day"))
v1
}
