% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pairwise_clust.R
\name{dist_pairwise_clust}
\alias{dist_pairwise_clust}
\title{compute distances and groups from algorithm based on raw distributions}
\usage{
dist_pairwise_clust(
  .data,
  gran1 = NULL,
  gran2 = NULL,
  response = NULL,
  quantile_prob_val = seq(0.1, 0.9, 0.1),
  kopt = NULL
)
}
\arguments{
\item{quantile_prob_val}{}
}
\value{

}
\description{
compute distances and groups from algorithm based on raw distributions
}
\examples{
library(gravitas)
library(tidyverse)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "day_month"
gran2 = NULL
response = "general_supply_kwh"
v2 <- suppressWarnings(dist_pairwise_clust(sm, "hour_day", quantile_prob_val = c(0.25, 0.5, 0.75)))
v2
}
