% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_wpd.R
\name{dist_wpd}
\alias{dist_wpd}
\title{Title}
\usage{
dist_wpd(.data, harmony_tbl = NULL, filter_comb = NULL, nperm = 2, nsamp = 2)
}
\arguments{
\item{nsamp}{}
}
\value{

}
\description{
Title
}
\examples{
library(gravitas)
library(tidyverse)
library(parallel)
library(tsibble)
library(rlang)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
  harmonies <- sm \%>\%
  harmony(
  ugran = "month",
  filter_in = "wknd_wday",
  filter_out = c("hhour", "fortnight")
  )
harmonies1 <- harmonies \%>\% mutate(facet_variable = NA)

 h = harmonies1 \%>\% select(-facet_levels) \%>\% distinct() \%>\% mutate(facet_levels = NA)
v = suppressWarnings(dist_wpd(sm, harmony_tbl = h))
v
}
