% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{robust_scaling of data marginal on categories of different granularities}
\usage{
plot_data(
  .data,
  gran1 = c,
  gran2 = NULL,
  response = NULL,
  quantile_prob_val = c(0.25, 0.5, 0.75),
  nfacet = NULL,
  method = "robust"
)
}
\arguments{
\item{method}{}
}
\value{
plot_data
}
\description{
robust_scaling of data marginal on categories of different granularities
}
\examples{
library(gravitas)
library(tidyverse)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
response = "general_supply_kwh"
v2 <- suppressWarnings(plot_data(sm, "hour_day"))
v2
sm <- robust_scale_data(sm, "hour_day")
v3 <- suppressWarnings(plot_data(sm, "hour_day", response = "scaled_response"))
#todo : can add NQT marginal on each category?
}
