% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_scale_data.R
\name{robust_scale_data}
\alias{robust_scale_data}
\title{robust_scaling of data marginal on categories of different granularities}
\usage{
robust_scale_data(
  .data,
  gran1 = c,
  gran2 = NULL,
  response = NULL,
  method = "robust"
)
}
\arguments{
\item{method}{}
}
\value{
robust_scale_data
}
\description{
robust_scaling of data marginal on categories of different granularities
}
\examples{
library(gravitas)
library(tidyverse)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
response = "general_supply_kwh"
v2 <- suppressWarnings(robust_scale_data(sm, "hour_day"))
v2
#todo : can add NQT marginal on each category?
}
