% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_dist_gran.R
\name{scaled_dist_gran}
\alias{scaled_dist_gran}
\title{compute distances and groups from algorithm based on raw distributions}
\usage{
scaled_dist_gran(
  .data,
  gran1 = NULL,
  gran2 = NULL,
  response = NULL,
  quantile_prob_val = seq(0.1, 0.9, 0.1),
  kopt = NULL
)
}
\arguments{
\item{quantile_prob_val}{}
}
\value{

}
\description{
compute distances and groups from algorithm based on raw distributions
}
\examples{
library(gravitas)
library(tidyverse)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
response = "general_supply_kwh"
v2 <- suppressWarnings(scaled_dist_gran(sm, "hour_day"))
v2
v3 <- suppressWarnings(scaled_dist_gran(sm, "hour_day"))
v3
}
