% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_clust.R
\name{wpd_clust}
\alias{wpd_clust}
\title{Title}
\usage{
wpd_clust(
  .data,
  harmony_tbl = NULL,
  filter_comb = NULL,
  nperm = 2,
  nsamp = 2,
  kopt = 2
)
}
\arguments{
\item{nsamp}{}
}
\value{

}
\description{
Title
}
\examples{
library(gravitas)
library(tidyverse)
library(parallel)
library(tsibble)
library(rlang)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10006704", "10017936","10006414", "10018250"))
gran1 = "hour_day"
gran2 = NULL
  harmonies <- sm \%>\%
  harmony(
  ugran = "month",
  filter_in = "wknd_wday",
  filter_out = c("hhour", "fortnight")
  )
v = suppressWarnings(wpd_clust(sm, harmony_tbl = harmonies))
v
}
